move get toolbar button name func
